+++ Changelog.md
@@ -1,3 +1,7 @@
+--------------------------Changelog for v1.9.11-beta2 of SecurityCraft--------------------------
+
+- Fix: Security sea boats are not invulnerable to damage types they should be invulnerable to
+
 --------------------------Changelog for v1.9.11-beta1 of SecurityCraft--------------------------
 
 - Change: Item rarities of the Admin Tool and Universal Block Reinforcer (Lvl 3) have been adjusted